+++ asset_allocation/strategies/pick_all.py
@@ -25,8 +25,10 @@ class PickAll(strategy.Strategy):
     def assign(
         self,
         date: datetime.date,
-        price_return: np.array,
+        price_return: np.ndarray,
+        index_comp: np.ndarray,
         annualize_factor: int = 1.0,
+        **kwargs,
     ) -> None:
         """
         Transform and assign returns to the actual calculator
@@ -37,10 +39,12 @@ class PickAll(strategy.Strategy):
             date of snapshot
         price_return: np.array
             zero base price return of universe
+        index_comp: np.array
+            index components for date
         annualize_factor: int, optional
             factor depending on data frequency
         """
-        super().assign(date, price_return, annualize_factor)
+        super().assign(date, price_return, index_comp, annualize_factor)
 
         self.return_engine.assign(
             date=date, price_return=price_return, annualize_factor=annualize_factor
@@ -58,7 +62,7 @@ class PickAll(strategy.Strategy):
             )
 
     @property
-    def selected_securities(self) -> np.array:
+    def selected_securities(self) -> np.ndarray:
         """
         Index (position in universe_tickers as integer) of selected securities
 
@@ -71,7 +75,7 @@ class PickAll(strategy.Strategy):
         return ss[~np.isnan(self.latest_return) & self.index_comp]
 
     @property
-    def return_metrics_optimizer(self) -> np.array:
+    def return_metrics_optimizer(self) -> np.ndarray:
         """
         Forecaseted DAILY returns from return engine in order of selected_securities