+++ CHANGELOG.md
@@ -11,6 +11,27 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 ### Removed
 
+## [1.5.0] -2023-12-05
+### Added
+- Documentation for Risk Framework, PAI and Asset Allocation
+- Relative Value Strategy
+- Market Multiple Datasource
+- Correlation test in test file
+- Hierarchical Risk Parity allocator
+- TCW Equity Portfolio Adjustments
+### Fixed
+- MSCI Datasource empty return
+- Numpy array value specification to np.ndarray
+- Momentum top_n bug
+### Changed
+- README
+- Snowflake pull function uses either query or database/schema/tablename
+- MSCI Dataloader allow specified factors
+- Specific Configs files for risk_framework, asset_allocation, pai
+- Move main runner into runners folder
+- empty list portfolios = load nothing, ["all] = load all
+### Removed
+
 ## [1.4.0] - 2023-11-24
 ### Added
 - JPM CEMBI BROAD DIVERSE and PM EMBI GLOBAL DIVERSIFI added to portfolio df
@@ -191,6 +212,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 
 [unreleased]: https://gitlab.com/tcw-group/quant-research/quantkit/-/compare/main...develop?from_project_id=46798372&straight=false
+[1.5.0]: https://gitlab.com/tcw-group/quant-research/quantkit/-/releases/1.4.0
 [1.4.0]: https://gitlab.com/tcw-group/quant-research/quantkit/-/releases/1.4.0
 [1.3.3]: https://gitlab.com/tcw-group/quant-research/quantkit/-/releases/1.3.3
 [1.3.2]: https://gitlab.com/tcw-group/quant-research/quantkit/-/releases/1.3.2